Repository refactor

+++ app/src/main/java/com/cassnyo/brasero/data/repository/TownForecastRepository.kt
@@ -7,48 +7,54 @@ import com.cassnyo.brasero.data.network.AemetApi
 import com.cassnyo.brasero.data.network.response.common.ForecastApi
 import com.cassnyo.brasero.data.network.response.daily.DailyForecastItemDayApi
 import com.cassnyo.brasero.data.network.response.hourly.HourlyForecastItemDayApi
+import com.cassnyo.brasero.di.module.IoDispatcher
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.withContext
+import java.time.LocalDateTime
 import javax.inject.Inject
 
 class TownForecastRepository @Inject constructor(
     private val aemetApi: AemetApi,
-    private val braseroDatabase: BraseroDatabase
+    private val braseroDatabase: BraseroDatabase,
+    @IoDispatcher private val ioDispatcher: CoroutineDispatcher
 ) {
 
-    fun getFavoriteTownsForecast(): Flow<List<TownForecast>> {
-        return braseroDatabase.townForecastDao().getFavoriteTownsForecast()
+    fun observeTownForecast(townId: String): Flow<TownForecast> {
+        return braseroDatabase.townForecastDao()
+            .getTownForecastById(townId)
+            .flowOn(ioDispatcher)
     }
 
-    fun getTownForecast(townId: String): Flow<TownForecast> {
-        return braseroDatabase.townForecastDao().getTownForecastById(townId)
-    }
+    suspend fun refreshTownForecast(
+        townId: String,
+        forceRefresh: Boolean = false
+    ) = withContext(ioDispatcher) {
+        val town = braseroDatabase.townDao().getTown(townId)
+        if (town == null || town.needsToUpdate() || forceRefresh) {
+            val dailyForecast = getRemoteDailyForecastByTown(townId).toDailyEntities(townId)
+            val hourlyForecast = getRemoteHourlyForecastByTown(townId).toHourlyEntities(townId)
 
-    suspend fun refreshTownForecast(townId: String) {
-        // TODO Observe Room query for a single town
-        // TODO refresh daily forecast
-        // TODO refresh hourly forecast
-        // TODO execute in parallel
-        val dailyForecast = getDailyForecastByTown(townId)
-        val hourlyForecast = getHourlyForecastByTown(townId)
-
-        val dailyForecastEntities = mapDailyForecastToEntities(dailyForecast, townId)
-        val hourlyForecastEntities = mapHourlyForecastToEntities(hourlyForecast, townId)
-
-        with(braseroDatabase) {
-            runInTransaction {
-                dayForecastDao().deleteDailyForecastByTownId(townId)
-                dayForecastDao().saveDailyForecast(dailyForecastEntities)
-                hourForecastDao().deleteHourlyForecastByTownId(townId)
-                hourForecastDao().saveHourlyForecast(hourlyForecastEntities)
+            with(braseroDatabase) {
+                runInTransaction {
+                    dayForecastDao().deleteDailyForecastByTownId(townId)
+                    dayForecastDao().saveDailyForecast(dailyForecast)
+                    hourForecastDao().deleteHourlyForecastByTownId(townId)
+                    hourForecastDao().saveHourlyForecast(hourlyForecast)
+                }
             }
         }
     }
 
-    private fun mapDailyForecastToEntities(
-        dailyForecast: ForecastApi<DailyForecastItemDayApi>,
-        townId: String
-    ): List<DayForecast> {
-        return dailyForecast.forecast.day.map { dayApi ->
+    private fun Town.needsToUpdate(): Boolean {
+        if (updatedAt == null) return true
+        val now = LocalDateTime.now()
+        return now.isAfter(updatedAt.plusHours(1))
+    }
+
+    private fun ForecastApi<DailyForecastItemDayApi>.toDailyEntities(townId: String): List<DayForecast> {
+        return forecast.day.map { dayApi ->
             // Group sky status by its description (using description as key) and then get the most repeated status
             val skyStatus = dayApi.skyStatus.groupBy { it.value }.maxByOrNull { it.value.size }?.key.orEmpty()
             // Use highest chanceOfRain of the whole day
@@ -78,22 +84,19 @@ class TownForecastRepository @Inject constructor(
         }
     }
 
-    private fun mapHourlyForecastToEntities(
-        hourlyForecast: ForecastApi<HourlyForecastItemDayApi>,
-        townId: String
-    ): List<HourForecast> {
+    private fun ForecastApi<HourlyForecastItemDayApi>.toHourlyEntities(townId: String): List<HourForecast> {
         val hourForecastMap = mutableListOf<HourForecast>()
 
-        hourlyForecast.forecast.day.forEach { day ->
+        forecast.day.forEach { day ->
 
-            (0 .. 24).forEach { hour ->
+            (0..24).forEach { hour ->
                 val fixedHour = when {
                     hour < 10 -> "0$hour"
                     else -> hour.toString().take(2)
                 }
 
                 val skyStatus = day.skyStatus.firstOrNull { it.period == fixedHour }
-                val chanceOfRain = day.chanceOfRain.firstOrNull { it.period?.take(2) == fixedHour}
+                val chanceOfRain = day.chanceOfRain.firstOrNull { it.period?.take(2) == fixedHour }
                 val rain = day.rain.firstOrNull { it.period == fixedHour }
                 val temperature = day.temperature.firstOrNull { it.period == fixedHour }
                 val wind = day.wind.firstOrNull { it.period == fixedHour }
@@ -104,7 +107,8 @@ class TownForecastRepository @Inject constructor(
                     rain != null ||
                     temperature != null ||
                     wind != null ||
-                    humidity != null) {
+                    humidity != null
+                ) {
                     val rainValue = try {
                         rain?.value?.toInt() ?: 0
                     } catch (e: NumberFormatException) {
@@ -137,12 +141,12 @@ class TownForecastRepository @Inject constructor(
         return hourForecastMap
     }
 
-    private suspend fun getDailyForecastByTown(townId: String): ForecastApi<DailyForecastItemDayApi> {
+    private suspend fun getRemoteDailyForecastByTown(townId: String): ForecastApi<DailyForecastItemDayApi> {
         val wrapper = aemetApi.getDailyForecastByTownWrapper(townId)
         return aemetApi.getDailyForecastByTown(wrapper.data).first()
     }
 
-    private suspend fun getHourlyForecastByTown(townId: String): ForecastApi<HourlyForecastItemDayApi> {
+    private suspend fun getRemoteHourlyForecastByTown(townId: String): ForecastApi<HourlyForecastItemDayApi> {
         val wrapper = aemetApi.getHourlyForecastByTownWrapper(townId)
         return aemetApi.getHourlyForecastByTown(wrapper.data).first()
     }
